Changes necessary for macOS to compile.
macOS differentiates between size_t (unsigned long long int) and uint64_t (unsigned long int)

This will cause compile issues.

Changed MeshIndexType to size_t.

+++ src/complex/DataStructure/Geometry/AbstractGeometry.hpp
@@ -111,7 +111,7 @@ public:
     Unknown = 101U
   };
 
-  using MeshIndexType = uint64_t;
+  using MeshIndexType = size_t;
   using MeshIndexArrayType = DataArray<MeshIndexType>;
   using SharedVertexList = FloatArray;
   using SharedEdgeList = MeshIndexArrayType;
